TGP-1267/updateRecord: refactoring
