+++ app/uk/gov/hmrc/tradergoodsprofilesrouter/controllers/UpdateRecordController.scala
@@ -24,10 +24,10 @@ import play.api.libs.json.{JsValue, Json}
 import play.api.mvc.{Action, ControllerComponents, Request, Result}
 import uk.gov.hmrc.play.bootstrap.backend.controller.BackendBaseController
 import uk.gov.hmrc.tradergoodsprofilesrouter.controllers.action.ValidationRules
+import uk.gov.hmrc.tradergoodsprofilesrouter.controllers.action.ValidationRules.optionalFieldsToErrorCode
 import uk.gov.hmrc.tradergoodsprofilesrouter.models.request.UpdateRecordRequest
 import uk.gov.hmrc.tradergoodsprofilesrouter.models.response.errors.BadRequestErrorResponse
 import uk.gov.hmrc.tradergoodsprofilesrouter.service.{UpdateRecordService, UuidService}
-import uk.gov.hmrc.tradergoodsprofilesrouter.utils.ValidationSupport.optionalFieldsToErrorCode
 
 import scala.concurrent.{ExecutionContext, Future}
