+++ app/uk/gov/hmrc/tradergoodsprofilesrouter/controllers/action/ValidationRules.scala
@@ -19,17 +19,22 @@ package uk.gov.hmrc.tradergoodsprofilesrouter.controllers.action
 import cats.data.EitherT
 import cats.implicits.catsSyntaxTuple2Parallel
 import cats.syntax.all._
-import play.api.libs.json.{JsValue, Reads}
+import org.apache.commons.validator.routines.EmailValidator
+import play.api.libs.functional.syntax.toApplicativeOps
+import play.api.libs.json.Reads.{maxLength, minLength, verifying}
+import play.api.libs.json._
 import play.api.mvc.{BaseController, Request, Result}
-import uk.gov.hmrc.tradergoodsprofilesrouter.controllers.action.ValidationRules.ValidatedQueryParameters
+import uk.gov.hmrc.tradergoodsprofilesrouter.controllers.action.ValidationRules.{ValidatedQueryParameters, extractSimplePaths, isValidActorId}
 import uk.gov.hmrc.tradergoodsprofilesrouter.models.response.errors.{BadRequestErrorResponse, Error}
 import uk.gov.hmrc.tradergoodsprofilesrouter.service.UuidService
 import uk.gov.hmrc.tradergoodsprofilesrouter.utils.ApplicationConstants._
-import uk.gov.hmrc.tradergoodsprofilesrouter.utils.{HeaderNames, ValidationSupport}
+import uk.gov.hmrc.tradergoodsprofilesrouter.utils.HeaderNames
 
-import java.util.UUID
+import java.time.Instant
+import java.util.{Locale, UUID}
 import scala.concurrent.{ExecutionContext, Future}
 import scala.util.Try
+import scala.util.matching.Regex
 
 //todo: we may want to unify ValidationSupport and this one.
 trait ValidationRules {
@@ -56,18 +61,16 @@ trait ValidationRules {
       )
     )
 
-  protected def validateActorId(actorId: String): Either[Error, String] = {
-    val pattern = "^[A-Z]{2}\\d{12,15}$".r
-    pattern
-      .findFirstIn(actorId)
-      .toRight(
+  protected def validateActorId(actorId: String): Either[Error, String] =
+    if (isValidActorId(actorId)) Right(actorId)
+    else
+      Left(
         Error(
           InvalidQueryParameter,
           InvalidActorIdQueryParameter,
           InvalidOrMissingActorIdCode.toInt
         )
       )
-  }
 
   protected def validateRequestBody[A: Reads](
     fieldToErrorCodeTable: Map[String, (String, String)]
@@ -81,11 +84,11 @@ trait ValidationRules {
       .leftMap { errors =>
         BadRequestErrorResponse(
           uuidService.uuid,
-          ValidationSupport.convertError[A](errors, fieldToErrorCodeTable)
+          convertError[A](errors, fieldToErrorCodeTable)
         ).asPresentation
       }
 
-  def validateQueryParameters(actorId: String, recordId: String) =
+  protected def validateQueryParameters(actorId: String, recordId: String) =
     (
       validateActorId(actorId).toEitherNec,
       validateRecordId(recordId).toEitherNec
@@ -94,9 +97,104 @@ trait ValidationRules {
     } leftMap { errors =>
       errors.toList
     }
+
+  private def convertError[T](
+    errors: scala.collection.Seq[(JsPath, scala.collection.Seq[JsonValidationError])],
+    fieldToErrorCodeTable: Map[String, (String, String)]
+  ): Seq[Error] =
+    extractSimplePaths(errors)
+      .map(key => fieldToErrorCodeTable.get(key).map(res => Error.invalidRequestParameterError(res._2, res._1.toInt)))
+      .toSeq
+      .flatten
 }
 
 object ValidationRules {
 
   final case class ValidatedQueryParameters(actorId: String, recordId: String)
+
+  val actorIdPattern: Regex = raw"[A-Z]{2}\d{12,15}".r
+  val comcodePattern: Regex = raw".{6}(.{2}(.{2})?)?".r
+
+  def isValidCountryCode(rawCountryCode: String): Boolean =
+    Locale.getISOCountries.toSeq.contains(rawCountryCode.toUpperCase)
+
+  def isValidDate(instant: Instant): Boolean =
+    instant.getNano == 0
+
+  private val emailValidator: EmailValidator = EmailValidator.getInstance(true)
+
+  object Reads {
+    def lengthBetween(min: Int, max: Int): Reads[String] =
+      minLength[String](min).keepAnd(maxLength[String](max))
+
+    val validActorId: Reads[String] = verifying(isValidActorId)
+
+    val validComcode: Reads[String] = verifying(isValidComcode)
+
+    val validDate: Reads[Instant]        = verifying(isValidDate)
+    val validEmailAddress: Reads[String] = verifying(isValidEmailAddress)
+  }
+
+  def isValidEmailAddress(emailAddress: String): Boolean = emailValidator.isValid(emailAddress)
+
+  def isValidActorId(actorId: String): Boolean = actorIdPattern.matches(actorId)
+
+  def isValidComcode(comcode: String): Boolean = comcodePattern.matches(comcode)
+
+  private def extractSimplePaths(
+    errors: scala.collection.Seq[(JsPath, collection.Seq[JsonValidationError])]
+  ): collection.Seq[String] =
+    errors
+      .map(_._1)
+      .map(_.path.filter(_.isInstanceOf[KeyPathNode]))
+      .map(_.mkString)
+
+  val fieldsToErrorCode: Map[String, (String, String)] = Map(
+    "/eori"                                       -> (InvalidOrMissingEoriCode, InvalidOrMissingEori),
+    "/recordId"                                   -> (RecordIdDoesNotExistsCode, InvalidRecordId),
+    "/actorId"                                    -> (InvalidOrMissingActorIdCode, InvalidOrMissingActorId),
+    "/traderRef"                                  -> (InvalidOrMissingTraderRefCode, InvalidOrMissingTraderRef),
+    "/comcode"                                    -> (InvalidOrMissingComcodeCode, InvalidOrMissingComcode),
+    "/goodsDescription"                           -> (InvalidOrMissingGoodsDescriptionCode, InvalidOrMissingGoodsDescription),
+    "/countryOfOrigin"                            -> (InvalidOrMissingCountryOfOriginCode, InvalidOrMissingCountryOfOrigin),
+    "/category"                                   -> (InvalidOrMissingCategoryCode, InvalidOrMissingCategory),
+    "/assessments/assessmentId"                   -> (InvalidOrMissingAssessmentIdCode, InvalidOrMissingAssessmentId),
+    "/assessments/primaryCategory"                -> (InvalidAssessmentPrimaryCategoryCode, InvalidAssessmentPrimaryCategory),
+    "/assessments/condition/type"                 -> (InvalidAssessmentPrimaryCategoryConditionTypeCode, InvalidAssessmentPrimaryCategoryConditionType),
+    "/assessments/condition/conditionId"          -> (InvalidAssessmentPrimaryCategoryConditionIdCode, InvalidAssessmentPrimaryCategoryConditionId),
+    "/assessments/condition/conditionDescription" -> (InvalidAssessmentPrimaryCategoryConditionDescriptionCode, InvalidAssessmentPrimaryCategoryConditionDescription),
+    "/assessments/condition/conditionTraderText"  -> (InvalidAssessmentPrimaryCategoryConditionTraderTextCode, InvalidAssessmentPrimaryCategoryConditionTraderText),
+    "/supplementaryUnit"                          -> (InvalidOrMissingSupplementaryUnitCode, InvalidOrMissingSupplementaryUnit),
+    "/measurementUnit"                            -> (InvalidOrMissingMeasurementUnitCode, InvalidOrMissingMeasurementUnit),
+    "/comcodeEffectiveFromDate"                   -> (InvalidOrMissingComcodeEffectiveFromDateCode, InvalidOrMissingComcodeEffectiveFromDate),
+    "/comcodeEffectiveToDate"                     -> (InvalidOrMissingComcodeEffectiveToDateCode, InvalidOrMissingComcodeEffectiveToDate),
+    "/requestorName"                              -> (InvalidMissingRequestorNameCode, InvalidOrMissingRequestorName),
+    "/requestorEmail"                             -> (InvalidMissingRequestorEmailCode, InvalidOrMissingRequestorEmail),
+    "/ukimsNumber"                                -> (InvalidOrMissingUkimsNumberCode, InvalidOrMissingUkimsNumberMessage),
+    "/nirmsNumber"                                -> (InvalidOrMissingNirmsNumberCode, InvalidOrMissingNirmsNumberMessage),
+    "/niphlNumber"                                -> (InvalidOrMissingNiphlNumberCode, InvalidOrMissingNiphlNumberMessage)
+  )
+
+  val optionalFieldsToErrorCode: Map[String, (String, String)] = Map(
+    "/eori"                                       -> (InvalidOrMissingEoriCode, InvalidOrMissingEori),
+    "/recordId"                                   -> (RecordIdDoesNotExistsCode, InvalidRecordId),
+    "/actorId"                                    -> (InvalidOrMissingActorIdCode, InvalidOrMissingActorId),
+    "/traderRef"                                  -> (InvalidOrMissingTraderRefCode, InvalidOrMissingOptionalTraderRef),
+    "/comcode"                                    -> (InvalidOrMissingComcodeCode, InvalidOrMissingOptionalComcode),
+    "/goodsDescription"                           -> (InvalidOrMissingGoodsDescriptionCode, InvalidOrMissingOptionalGoodsDescription),
+    "/countryOfOrigin"                            -> (InvalidOrMissingCountryOfOriginCode, InvalidOrMissingOptionalCountryOfOrigin),
+    "/category"                                   -> (InvalidOrMissingCategoryCode, InvalidOrMissingOptionalCategory),
+    "/assessments/assessmentId"                   -> (InvalidOrMissingAssessmentIdCode, InvalidOrMissingAssessmentId),
+    "/assessments/primaryCategory"                -> (InvalidAssessmentPrimaryCategoryCode, InvalidAssessmentPrimaryCategory),
+    "/assessments/condition/type"                 -> (InvalidAssessmentPrimaryCategoryConditionTypeCode, InvalidAssessmentPrimaryCategoryConditionType),
+    "/assessments/condition/conditionId"          -> (InvalidAssessmentPrimaryCategoryConditionIdCode, InvalidAssessmentPrimaryCategoryConditionId),
+    "/assessments/condition/conditionDescription" -> (InvalidAssessmentPrimaryCategoryConditionDescriptionCode, InvalidAssessmentPrimaryCategoryConditionDescription),
+    "/assessments/condition/conditionTraderText"  -> (InvalidAssessmentPrimaryCategoryConditionTraderTextCode, InvalidAssessmentPrimaryCategoryConditionTraderText),
+    "/supplementaryUnit"                          -> (InvalidOrMissingSupplementaryUnitCode, InvalidOrMissingSupplementaryUnit),
+    "/measurementUnit"                            -> (InvalidOrMissingMeasurementUnitCode, InvalidOrMissingMeasurementUnit),
+    "/comcodeEffectiveFromDate"                   -> (InvalidOrMissingComcodeEffectiveFromDateCode, InvalidOrMissingOptionalComcodeEffectiveFromDate),
+    "/comcodeEffectiveToDate"                     -> (InvalidOrMissingComcodeEffectiveToDateCode, InvalidOrMissingComcodeEffectiveToDate),
+    "/requestorName"                              -> (InvalidOrMissingRequestorNameCode, InvalidOrMissingRequestorName),
+    "/requestorEmail"                             -> (InvalidOrMissingRequestorEmailCode, InvalidOrMissingRequestorEmail)
+  )
 }
