+++ app/uk/gov/hmrc/tradergoodsprofilesrouter/models/request/MaintainProfileRequest.scala
@@ -19,7 +19,7 @@ package uk.gov.hmrc.tradergoodsprofilesrouter.models.request
 import play.api.libs.functional.syntax.toFunctionalBuilderOps
 import play.api.libs.json.Format.GenericFormat
 import play.api.libs.json.{JsPath, OWrites, Reads}
-import uk.gov.hmrc.tradergoodsprofilesrouter.utils.ValidationSupport.Reads.{lengthBetween, validActorId}
+import uk.gov.hmrc.tradergoodsprofilesrouter.controllers.action.ValidationRules.Reads.{lengthBetween, validActorId}
 
 import scala.Function.unlift
